pyproject.toml
[tool.poetry]
name = "esp32-configurator"
version = "2.1.0"
description = "ESP32 devices and sensors configurator"
authors = ["Stefano Cappa"]
readme = "README.md"
packages = [
    {include = "src"}
]

[tool.poetry.scripts]
esp32-configurator = "src.__main__:main"

[tool.poetry.dependencies]
python = "^3.12"
jinja2 = "^3.1.6"
pydantic = "^2.12.5"
pyyaml = "^6.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
